feat(examples): retorrent
A CLI tool to reload torrents based on params.

The same params are used to query torrents from API.

This tool then delete the torrent from client, use `mktorrent` to build
a brand new torrent file, and add that to the client again.

We have to rebuild the torrent file because we can't add using magnet
URI for private trackers.

See code for options, including `no-check`.